+++ src/biodwh2-core/src/main/java/de/unibi/agbi/biodwh2/core/model/graph/NodeMappingDescription.java
@@ -27,6 +27,7 @@ public final class NodeMappingDescription {
         PHYTOCHEMICAL,
         PRODRUG,
         PROTEIN,
+        PROTEIN_DOMAIN,
         PUBLICATION,
         RNA,
         //SIDE_EFFECT, //-------