+++ src/biodwh2-negatome/src/main/java/de/unibi/agbi/biodwh2/negatome/model/PfamPair.java
@@ -1,0 +1,8 @@
+package de.unibi.agbi.biodwh2.negatome.model;
+
+public class PfamPair {
+    public String pfamId1;
+    public String pfamId2;
+    public boolean isManual;
+    public boolean isPDB;
+}